Move unused snippets to closet.py for possible later use
